+++ test/unit/rx-database.test.ts
@@ -19,10 +19,12 @@ import {
     humansCollection,
     getPassword
 } from '../../plugins/test-utils/index.mjs';
-
+import {
+    getRxStorageMemory
+} from '../../plugins/storage-memory/index.mjs';
 
 describeParallel('rx-database.test.ts', () => {
-    describe('.create()', () => {
+    describe('createRxDatabase()', () => {
         describe('positive', () => {
             it('normal', async () => {
                 const db = await createRxDatabase({
@@ -41,7 +43,7 @@ describeParallel('rx-database.test.ts', () => {
                 assert.ok(isRxDatabase(db));
                 db.destroy();
             });
-            it('2 instances on same adapter (if ignoreDuplicate is true)', async () => {
+            it('2 instances on same storage (if ignoreDuplicate is true)', async () => {
                 const name = randomCouchString(10);
                 const db = await createRxDatabase({
                     name,
@@ -88,6 +90,24 @@ describeParallel('rx-database.test.ts', () => {
                     name,
                     storage: config.storage.getStorage()
                 });
+                db2.destroy();
+            });
+            it('do allow 2 databases with same name but different storage', async () => {
+                if (
+                    config.storage.name === 'memory'
+                ) {
+                    return;
+                }
+                const name = randomCouchString(10);
+                const db = await createRxDatabase({
+                    name,
+                    storage: getRxStorageMemory()
+                });
+                const db2 = await createRxDatabase({
+                    name,
+                    storage: config.storage.getStorage()
+                });
+                db.destroy();
                 db2.destroy();
             });
             it('2 password-instances on same adapter', async () => {
@@ -194,7 +214,7 @@ describeParallel('rx-database.test.ts', () => {
                     'ending'
                 );
             });
-            it('do not allow 2 databases with same name and adapter', async () => {
+            it('do not allow 2 databases with same name and storage', async () => {
                 const name = randomCouchString(10);
                 const db = await createRxDatabase({
                     name,